pyproject.toml
[project]
name = "devvdxwthybot"
version = "0.1.0"
description = ""
authors = [
    {name = "vdxwthy",email = "e.mikhaylov2006@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram (>=3.21.0,<4.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
]

[tool.poetry]
packages = [{include = "devvdxwthybot", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
